// client/src/types/auth.ts
export interface LoginRequest {
    email: string;
    password: string;
}
// ...
export interface AuthResponse {
    message: string;
    token?: string;
    tokenExpiresAt?: string;
    displayName?: string;
    avatarUrl?: string;
    role?: string;
    identityUserId?: string;
    email?: string;
    workStatus?: number;
}